refactor: create runner/names package
